+++ examples/marvel/lib/src/characters.json
