+++ examples/marvel/lib/src/configuration.dart
