+++ examples/marvel/lib/src/fake_marvel.dart
