+++ examples/marvel/lib/src/marvel.dart
