+++ examples/marvel/lib/src/result.dart
